+++ CHANGELOG.md
@@ -5,8 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
 ### Changed
+- Python typing features added to all modules
+## [0.8] - 2025-12-17
+### Added
+- `WinMM` engine
+### Changed
+- `Python 3.14` added to tests
+- `run` function in `NavaThread` updated
+- `stop` function in `NavaThread` updated
 - `README.md` updated
 - Test system modified
 - `Python 3.6` support dropped
@@ -75,7 +82,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `play` function
 
 
-[Unreleased]: https://github.com/openscilab/nava/compare/v0.7...dev
+[Unreleased]: https://github.com/openscilab/nava/compare/v0.8...dev
+[0.8]: https://github.com/openscilab/nava/compare/v0.7...v0.8
 [0.7]: https://github.com/openscilab/nava/compare/v0.6...v0.7
 [0.6]: https://github.com/openscilab/nava/compare/v0.5...v0.6
 [0.5]: https://github.com/openscilab/nava/compare/v0.4...v0.5